Alternative fix: promote cycleslate to an s64 everywhere.
Also changed a few functions to be static.

+++ Source/Core/Core/CoreTiming.h
@@ -35,7 +35,7 @@ extern float g_lastOCFactor_inverted;
 void Init();
 void Shutdown();
 
-typedef void (*TimedCallback)(u64 userdata, int cyclesLate);
+typedef void (*TimedCallback)(u64 userdata, s64 cyclesLate);
 
 // This should only be called from the CPU thread, if you are calling it any other thread, you are doing something evil
 u64 GetTicks();
